refactor: split server stats api client
